Refactor SSM benchmark & fixture test
