+++ monad-bayes.cabal
@@ -167,6 +167,7 @@ test-suite monad-bayes-test
     LDA
     LogReg
     NonlinearSSM
+    NonlinearSSM.Algorithms
     Sprinkler
     TestAdvanced
     TestBenchmarks
@@ -205,7 +206,10 @@ benchmark ssm-bench
   type:             exitcode-stdio-1.0
   main-is:          SSM.hs
   hs-source-dirs:   models benchmark
-  other-modules:    NonlinearSSM
+  other-modules:
+    NonlinearSSM
+    NonlinearSSM.Algorithms
+
   default-language: Haskell2010
   build-depends:
     , base
